added tests for skip policy and some refactor
