+++ eva-accession-release/src/main/java/uk/ac/ebi/eva/accession/release/exceptions/IllegalStartPositionException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.release;
+package uk.ac.ebi.eva.accession.release.exceptions;
 
 public class IllegalStartPositionException extends IllegalArgumentException {
